feature/game-enhancements-q3-2025
fix: Correct non-functional back button on options screen

This commit fixes a bug where the 'Back' button on the Options screen would not work when the screen was accessed from the Main Menu.

The issue was caused by a subtle problem in the state of the `source_screen` navigation variable. To make the navigation more robust and to immediately solve the issue, the `on_press` event for the back button in `optionsscreen.kv` has been changed to call `app.change_screen('main')` directly, instead of relying on the `go_back()` method.